// services/cCREs-service/graphqlservice/src/postgres/types.ts
export type SCREENSearchParameters = {
    accessions?: string[];
    assembly: string;
    cellType?: string;
    coord_chrom: string;
    coord_end: number;
    coord_start: number;
    element_type?: string;
    gene_all_start?: number;
    gene_all_end?: number;
    gene_pc_start?: number;
    gene_pc_end?: number;
    rank_ctcf_end: number;
    rank_ctcf_start: number;
    rank_dnase_end: number;
    rank_dnase_start: number;
    rank_enhancer_end: number;
    rank_enhancer_start: number;
    rank_promoter_end: number;
    rank_promoter_start: number;
    uuid: string;
};

export type SCREENSearchResult = {
    chrom: string;
    start: number;
    len: number;
    pct: string;
    ctcf_zscore: number;
    dnase_zscore: number;
    enhancer_zscore: number;
    promoter_zscore: number;
    genesallpc: {
      accession: string;
      all: {
        __typename: "IntersectingGenes";
        assembly: string;
        chromosome: string;
        start: number;
        end: number;
        limit: number;
        protein_coding?: boolean;
      };
      pc: {
        __typename: "IntersectingGenes";
        assembly: string;
        chromosome: string;
        start: number;
        end: number;
        limit: number;
        protein_coding: boolean;
      };
    };
    info: {
      accession: string;
      isproximal: boolean;
      concordant: boolean;
      ctcfmax: number;
      k4me3max: number;
      k27acmax: number;
    };
    vistaids?: string[];
    sct: number;
    maxz: number;
    in_cart: number;
};

export type GenomicRange = {
    chromosome: string;
    start: number;
    end: number;
};

export type BiosampleEntry = {
    biosample_name: string;
    [key: string]: string | null | undefined;
};

export type RDHSEntry = {
    accession: string;
    chromosome: string;
    start: number;
    stop: number;
    assembly?: string;
};

export type LinkedGenesParameters = {
    assembly?: string;
    accession: string[];
};

export type LinkedGene = {
    assembly: string;
    accession: string;
    assay: string;
    experiment_accession: string;
    celltype: string;
    gene: string;
};
// ...
export type CCREEntry = {
    accession: string;
    rdhs: string;
    ccre_group: string;
    ctcf_bound: boolean;
    chromosome: string;
    start: number;
    stop: number;
    assembly?: string;
};

export type CCREParameters = {
    assembly: string;
    accession_prefix?: string[];
    accession?: string[];
    rDHS?: string[];
    group?: string[];
    ctcf_bound?: boolean;
    coordinates?: GenomicRange[];
    limit?: number;
    activeInBiosamples?: string[];
    activeInAnyBiosample?: string[];
};

export type CCREMaxZParameters = {
    assembly: string;
    accession: string[];
    assay: string;
};

export type CCREMaxZEntry = {
    accession: string;
    score: number;
};

export type CellTypeAssayEntry = {
    biosample_name: string;
    [key: string]: string;
};

export type ZScoreEntry = {
    experiment_accession: string;
    rdhs: string;
    score: number;
};